Move state out of folder
