+++ packages/state/source/index.ts
@@ -1,4 +1,4 @@
 export * as Attributes from "./attributes";
 export * from "./service-provider";
-export * as Stores from "./stores";
+export * from "./state";
 export * as Wallets from "./wallets";
