+++ packages/state/source/service-provider.ts
@@ -6,8 +6,8 @@ import { AttributeRepository } from "./attributes";
 import { BlockState } from "./block-state";
 import { AttributeMutator } from "./mutators/attribute";
 import { BalanceMutator } from "./mutators/balance";
+import { StateStore } from "./state";
 import { StateVerifier } from "./state-verifier";
-import { StateStore } from "./stores/state";
 import { IndexSet, WalletRepository, WalletRepositoryClone, WalletRepositoryCopyOnWrite } from "./wallets";
 import { validatorWalletFactory, walletFactory } from "./wallets/factory";
