--- packages/state/source/stores/index.ts
@@ -1,1 +1,0 @@
-export * from "./state";
